Updated web/templates.go, extracted structs into web/template_structs.go
